refactor: changing archives paths and decimal_to_double function
