+++ lib/portfolio/repo/assets_repository_impl.dart
@@ -1,5 +1,14 @@
 import 'package:projeto_criptos/shared/models/asset_model.dart';
 
-abstract class IGetAssetsRepository {
-  List<AssetModel> getAssets();
+import '../../shared/repository/assets_repository.dart';
+
+
+class GetAssetsRepositoryImpl implements IGetAssetsRepository {
+  List<AssetModel> lista;
+
+  GetAssetsRepositoryImpl(this.lista);
+  @override
+  List<AssetModel> getAssets() {
+    return lista;
+  }
 }
