+++ lib/portfolio/widgets/asset_list_tile.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:projeto_criptos/portfolio/controller/visible_provider.dart';
-import 'package:projeto_criptos/shared/utils/decimal_parse.dart';
 
 import '../../shared/models/asset_model.dart';
 import '../../shared/utils/currency_formater.dart';
+import '../../shared/utils/decimal_to_double.dart';
 
 class AssetListTile extends HookConsumerWidget {
   const AssetListTile({
