+++ lib/shared/repository/assets_repository.dart
@@ -1,0 +1,5 @@
+import 'package:projeto_criptos/shared/models/asset_model.dart';
+
+abstract class IGetAssetsRepository {
+  List<AssetModel> getAssets();
+}
