+++ lib/shared/utils/decimal_to_double.dart
@@ -1,0 +1,4 @@
+//Decimal to double
+import 'package:decimal/decimal.dart';
+
+double dtd(Decimal source) => double.parse(source.toString());
